requirements.txt
# requirements.txt
streamlit
PyPDF2
spacy
networkx
sentence-transformers
scikit-learn
google-generativeai
python-dotenv
plotly
matplotlib
numpy

# Download spaCy model
&& python -m spacy download en_core_web_sm